fix: minor refactor rename of pulumi infra
