+++ apps/pulumi/index.ts
@@ -2,8 +2,8 @@ import * as pulumi from "@pulumi/pulumi";
 import * as gcp from "@pulumi/gcp";
 
 // Modules
-import { projectModule } from "./src.module/project";
-import { storageModule } from "./src.module/storage";
+import { projectModule } from "./src/project";
+import { storageModule } from "./src/storage";
 
 // Initialize project modules
 const project = projectModule();
